Watcher/requirements.txt
elasticsearch
discord.py
